chore: post-rebase updates
