+++ bc_obps/registration/api/__init__.py
@@ -1,10 +1,12 @@
-from . import contact
-from . import naics
-from . import operation
-from . import operator
-from . import user_operator
-from . import user
-from . import business_structure
-from . import regulated_products
-from . import reporting_activities
+from . import (
+    contact,
+    naics,
+    operation,
+    operator,
+    user_operator,
+    user,
+    business_structure,
+    regulated_products,
+    reporting_activities,
+)
 from .api_base import router  # django_ninja router object
