+++ bc_obps/registration/schema/__init__.py
@@ -5,3 +5,5 @@ from .operator import *
 from .user_operator import *
 from .user import *
 from .business_structure import *
+from .regulated_products import *
+from .reporting_activities import *
