+++ bc_obps/registration/schema/regulated_products.py
@@ -1,0 +1,12 @@
+from ninja import ModelSchema
+from registration.models import RegulatedProduct
+
+
+class RegulatedProductSchema(ModelSchema):
+    """
+    Schema for the RegulatedProduct model
+    """
+
+    class Config:
+        model = RegulatedProduct
+        model_fields = "__all__"
